requirements.txt
Django >= 4.2.1,<4.3
djangorestframework>=3.14.0,<3.15
psycopg2>=2.9,<2.10
drf-spectacular>=0.26.2,<0.27
djangorestframework-simplejwt>=5.2.2,<5.3
requests>=2.31.0,<2.32
django-ckeditor>=6.5.1,<6.6
gunicorn>=20.1.0,<20.2.0
django-cors-headers>=4.1.0,<4.2
Pillow>=10.0.0,<10.1
reportlab>=4.0.4,<4.1
django-admincharts>=0.4.1,<0.5
django-celery-beat>=2.5.0,<2.6
redis>=5.0.0,<5.1
django-elasticsearch-dsl>=7.3,<7.4
django-elasticsearch-dsl-drf>=0.22.5,<0.23